Create separate file for App class
